requirements.txt
Flask
pandas
requests
flask-cors
mysql-connector-python==8.0.29
